// src/app/shared/models/bicycleForCreateOrUpdateModel.ts
/**
 * BC.Bicycles
 * Microservice to manage bicycles
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface BicycleForCreateOrUpdateModel { 
    model: string;
    serialNumber?: string;
    userId: string;
    userName?: string;
}